Move closed popover overflow padding coverage to its own sandbox
The closed-popover fixture variant added to the combobox-overflow-padding
sandbox perturbed the page geometry that fixture's viewport-sized popover
and position-update counter depend on, making its reposition test flaky on
CI. Restore that sandbox to its original state and move the regression
coverage to a standalone popover-closed-overflow-padding sandbox. Also add
the perf CI quantification to both performance changesets.
